requirements.txt
Django==1.9.7
argparse==1.2.1
django-cors-headers==1.1.0
djangorestframework==3.4.0
wsgiref==0.1.2
